requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pypdf>=3.17.0
chromadb>=0.4.0